Refactor the code, move the ir to the onnx package, remove/rewrite torch legacy code, redesign the optimize/split APIs
