+++ onnx/array_ref.h
@@ -18,7 +18,7 @@
 #include <array>
 #include <vector>
 
-namespace onnx { namespace optimization {
+namespace onnx {
   /// ArrayRef - Represent a constant reference to an array (0 or more elements
   /// consecutively in memory), i.e. a start pointer and a length.  It allows
   /// various APIs to take consecutive elements easily and conveniently.
@@ -178,4 +178,4 @@ namespace onnx { namespace optimization {
     /// @}
   };
 
-}} // end namespace onnx::optimization
+} // namespace onnx
