+++ onnx/ir_pb_converter.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "onnx/ir.h"
+#include "onnx/onnx_pb.h"
+
+namespace onnx {
+
+void ExportModelProto(onnx::ModelProto* p_m, const std::shared_ptr<Graph>& g);
+std::unique_ptr<Graph> ImportModelProto(const onnx::ModelProto& mp);
+
+} // namespace onnx
