--- onnx/optimizer/utils/disallow_copy.h
@@ -1,4 +1,0 @@
-#pragma once
-#define TH_DISALLOW_COPY_AND_ASSIGN(TypeName) \
-  TypeName(const TypeName&) = delete; \
-  void operator=(const TypeName&) = delete
