+++ onnx/optimizer/wrapper.h
@@ -1,12 +1,11 @@
 #pragma once
 
+#include <list>
 #include <string>
 
 namespace onnx { namespace optimization {
 
-template <typename Proto>
-bool ParseProtoFromBytes(Proto* proto, const char* buffer, size_t length);
-
-std::string Optimize(const std::string& content, bool init, bool predict);
+std::string Optimize(const std::string& content, std::list<std::string>& names);
+std::string Split(const std::string& content, bool init, bool predict);
 
 }} // namespace onnx::optimization
